artifacts/api-server/ml/requirements.txt
flask==3.1.0
gunicorn==23.0.0
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
requests==2.32.3
scikit-learn==1.6.1
xgboost==2.1.3
